backend/requirements.txt
fastapi==0.65.2
psycopg2-binary==2.9.1
pydantic==1.8.2
SQLAlchemy==1.4.19
uvicorn==0.14.0
alembic==1.7.4
google-cloud-pubsub==2.13.6
firebase-admin==5.2.0
